+++ R/tm_t_binary_outcome.R
@@ -425,6 +425,9 @@ srv_t_binary_outcome <- function(input,
 
     anl_m <- anl_merged()
     input_aval_var <- as.vector(anl_m$columns_source$aval_var)
+    input_param <- paste0(gsub(paste0(unlist(anl_m$filter_info)["selected"], ": "), "",
+                               grep(unlist(anl_m$filter_info)["selected"], names(paramcd$filter[[1]]$choices), value = TRUE)),
+                          " (", unlist(anl_m$filter_info)["selected"], ")")
     req(input$responders %in% anl_m$data()[[input_aval_var]])
 
     chunks_push_data_merge(anl_m)
@@ -441,6 +444,7 @@ srv_t_binary_outcome <- function(input,
       dataname = "ANL",
       parentname = "ANL_ADSL",
       arm_var = as.vector(anl_m$columns_source$arm_var),
+      param = input_param,
       ref_arm = input$ref_arm,
       comp_arm = input$comp_arm,
       compare_arm = input$compare_arms,