+++ R/tm_t_rsp.R
@@ -41,6 +41,7 @@
 template_rsp <- function(dataname,
                          parentname,
                          arm_var,
+                         param = "Response",
                          ref_arm = NULL,
                          comp_arm = NULL,
                          compare_arm = FALSE,
@@ -136,7 +137,13 @@ template_rsp <- function(dataname,
   }
 
   layout_list <- list()
-  layout_list <- add_expr(layout_list, substitute(basic_table()))
+  layout_list <- add_expr(
+    layout_list,
+    substitute(
+      expr = basic_table(title = paste("Table of", param, "for", paste(head(responders, -1), collapse = ", "),
+                                       ifelse(length(responders) > 1, "and", ""), tail(responders, 1), "Responders")),
+      env = list(param = param, responders = responder_val))
+    )
   layout_list <- add_expr(
     layout_list,
     split_col_expr(
@@ -695,6 +702,9 @@ srv_t_rsp <- function(input,
 
     anl_m <- anl_merged()
     input_aval_var <- as.vector(anl_m$columns_source$aval_var)
+    input_param <- paste0(gsub(paste0(unlist(anl_m$filter_info)["selected"], ": "), "",
+                        grep(unlist(anl_m$filter_info)["selected"], names(paramcd$filter[[1]]$choices), value = TRUE)),
+                        " (", unlist(anl_m$filter_info)["selected"], ")")
     req(input$responders %in% anl_m$data()[[input_aval_var]])
 
     chunks_push_data_merge(anl_m)
@@ -713,6 +723,7 @@ srv_t_rsp <- function(input,
       dataname = "ANL",
       parentname = "ANL_ADSL",
       arm_var = as.vector(anl_m$columns_source$arm_var),
+      param = input_param,
       ref_arm = input$ref_arm,
       comp_arm = input$comp_arm,
       compare_arm = input$compare_arms,